+++ Sources/Events/EventsClient.swift
@@ -1,15 +1,6 @@
 import Foundation
 
 
-import Foundation
-
-class EventsClientFactory {
-    static func createEventsClient(projectId: String, sdkType: String, sdkVersion: String, bundleId: String) -> EventsClient {
-        let storage = UserDefaultsEventStorage()
-        let networkingService = NetworkingService(projectId: projectId, sdkType: sdkType, sdkVersion: sdkVersion)
-        return EventsClient(storage: storage, bundleId: bundleId, networkingService: networkingService)
-    }
-}
 class EventsClient {
     private let eventsCollector: EventsCollector
     private let eventsDispatcher: EventsDispatcher
